AURORA: For GFF4 reference deduplication, use offset and struct index
Instead of just the offset. The field offsets are what counts in the
end, and those are calculate using the struct offset and the struct
template field offsets. This opens up the possibility of running into
collisions, especially with the top level struct.

In fact, collisions happen with the ARL files describing the area
environments in Dragon Age: Origins.

+++ src/xml/gff4dumper.h
@@ -61,7 +61,7 @@ private:
 		GFF4Field(const Aurora::GFF4Struct &s, uint32 f, bool g);
 	};
 
-	typedef std::set<uint32> IDSet;
+	typedef std::set<uint64> IDSet;
 	typedef std::map<uint32, Common::UString> FieldNames;
 
 
@@ -75,7 +75,7 @@ private:
 	IDSet _structIDs;
 
 
-	bool insertID(uint32 id);
+	bool insertID(uint64 id);
 
 	Common::UString getIFieldTypeName(uint32 type, bool isList) const;
